+++ LevelHelper2-API/Utilities/SKNode+Transforms.m
@@ -39,20 +39,18 @@
     CGSize size = [self size];
     
     
-    float anchorPointX = size.width*[self anchor].x;
-    float anchorPointY = size.height*[self anchor].y;
+    CGFloat anchorPointX = size.width*[self anchor].x;
+    CGFloat anchorPointY = size.height*[self anchor].y;
     
-    float centerPointX = size.width*0.5;
-    float centerPointY = size.height*0.5;
+    CGFloat centerPointX = size.width*0.5;
+    CGFloat centerPointY = size.height*0.5;
     
     CGPoint pos = [self position];
-    float rot = [self zRotation];
+    CGFloat rot = [self zRotation];
     
-    float xScl = [self xScale];
-    float yScl = [self yScale];
+    CGFloat xScl = [self xScale];
+    CGFloat yScl = [self yScale];
     
-
-
     CGAffineTransform nodeTransform = CGAffineTransformMakeTranslation(pos.x, pos.y);
     
     nodeTransform = CGAffineTransformRotate(nodeTransform, rot);
